Split main.go file.
